backend/requirements.txt
Django>=4.0.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.1.0
django-filter>=23.0
dj-rest-auth>=4.0.0
django-allauth>=0.54.0
djangorestframework-simplejwt>=5.2.0
django-oauth-toolkit>=2.3.0
drf-yasg>=1.21.6
python-dotenv>=0.19.0
